Add quote method
This method uses a regular expression that removes characters up to the first A-Z, a-z, or 0-9.  I hope that no quotes would begin with a character outside of this range.  Hmm.  This had to be used because WikiQuote contained the character ￼ in the place before the quote (got parsed into the string), and this I feel is safer than adding special handling for just that character.  Though, maybe that is better?  Hmm.